tx: improved type documentation, added TxValuesArray type alias
